requirements.txt
pydantic-settings
langchain-core
langchain-community
langchain-openai
chromadb
openai
pyrogram
python-dotenv
requests
tiktoken